+++ controllers/front/validation.php
@@ -58,6 +58,10 @@ class CashwayValidationModuleFrontController extends ModuleFrontController
             Tools::redirect('index.php?controller=order&step=1');
         }
 
+        if (Tools::getValue('cgu-accept') === false) {
+            Tools::redirect('index.php?fc=module&module=cashway&controller=payment&cgu=1');
+        }
+
         $currency = $this->context->currency;
         $total = (float)$cart->getOrderTotal(true, Cart::BOTH);
 
@@ -69,7 +73,10 @@ class CashwayValidationModuleFrontController extends ModuleFrontController
             $this->context->cart,
             $this->context->customer,
             $this->context->language->iso_code,
-            $cw_currency[0]['iso_code']
+            $cw_currency[0]['iso_code'],
+            array(
+                'button_type' => Tools::getValue('btn', '?')
+            )
         );
 
         $cw_res = $cashway->openTransaction();